refactor: make PostFundingCreate naming match use
Gotcha got me.
Additional refactoring addresses complexity code smells.

TIS21-8840: Funding Types require Subtypes when available
