+++ generic-upload-api/pom.xml
@@ -6,7 +6,7 @@
 
   <groupId>com.transformuk.hee.tis</groupId>
   <artifactId>generic-upload-api</artifactId>
-  <version>1.19.1</version>
+  <version>1.19.2</version>
   <packaging>jar</packaging>
 
   <properties>
