+++ generic-upload-api/src/main/java/com/transformuk/hee/tis/genericupload/api/enumeration/FileType.java
@@ -5,14 +5,14 @@ import com.transformuk.hee.tis.genericupload.api.dto.AssessmentUpdateXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.AssessmentXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.CurriculumMembershipCreateXls;
 import com.transformuk.hee.tis.genericupload.api.dto.CurriculumMembershipUpdateXls;
-import com.transformuk.hee.tis.genericupload.api.dto.FundingUpdateXLS;
+import com.transformuk.hee.tis.genericupload.api.dto.PostFundingUpdateRow;
 import com.transformuk.hee.tis.genericupload.api.dto.PersonUpdateXls;
 import com.transformuk.hee.tis.genericupload.api.dto.PersonXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.PlacementDeleteXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.PlacementUpdateXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.PlacementXls;
 import com.transformuk.hee.tis.genericupload.api.dto.PostCreateXls;
-import com.transformuk.hee.tis.genericupload.api.dto.PostFundingUpdateXLS;
+import com.transformuk.hee.tis.genericupload.api.dto.PostFundingCreateRow;
 import com.transformuk.hee.tis.genericupload.api.dto.PostUpdateXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.ProgrammeMembershipUpdateXls;
 import com.transformuk.hee.tis.genericupload.api.dto.TemplateXLS;
@@ -21,7 +21,6 @@ import lombok.Getter;
 public enum FileType {
   ASSESSMENTS(AssessmentXLS.class),
   ASSESSMENTS_DELETE(AssessmentDeleteXLS.class),
-  FUNDING_UPDATE(FundingUpdateXLS.class),
   PEOPLE(PersonXLS.class),
   PEOPLE_UPDATE(PersonUpdateXls.class),
   PLACEMENTS(PlacementXls.class),
@@ -29,7 +28,8 @@ public enum FileType {
   PLACEMENTS_UPDATE(PlacementUpdateXLS.class),
   POSTS_CREATE(PostCreateXls.class),
   POSTS_UPDATE(PostUpdateXLS.class),
-  POSTS_FUNDING_UPDATE(PostFundingUpdateXLS.class),
+  POSTS_FUNDING_CREATE(PostFundingCreateRow.class),
+  POSTS_FUNDING_UPDATE(PostFundingUpdateRow.class),
   ASSESSMENTS_UPDATE(AssessmentUpdateXLS.class),
   PROGRAMME_MEMBERSHIP_UPDATE(ProgrammeMembershipUpdateXls.class),
   CURRICULUM_MEMBERSHIP_CREATE(CurriculumMembershipCreateXls.class),
