+++ generic-upload-service/pom.xml
@@ -12,7 +12,7 @@
 
   <groupId>com.transformuk.hee.tis</groupId>
   <artifactId>generic-upload-service</artifactId>
-  <version>1.32.0</version>
+  <version>1.32.1</version>
 
   <packaging>war</packaging>
 
@@ -140,7 +140,7 @@
     <dependency>
       <groupId>com.transformuk.hee.tis</groupId>
       <artifactId>generic-upload-api</artifactId>
-      <version>1.19.1</version>
+      <version>1.19.2</version>
     </dependency>
     <dependency>
       <groupId>uk.nhs.tis</groupId>
