+++ generic-upload-service/src/main/java/com/transformuk/hee/tis/genericupload/service/service/ScheduledUploadTask.java
@@ -9,14 +9,14 @@ import com.transformuk.hee.tis.genericupload.api.dto.AssessmentUpdateXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.AssessmentXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.CurriculumMembershipCreateXls;
 import com.transformuk.hee.tis.genericupload.api.dto.CurriculumMembershipUpdateXls;
-import com.transformuk.hee.tis.genericupload.api.dto.FundingUpdateXLS;
+import com.transformuk.hee.tis.genericupload.api.dto.PostFundingUpdateRow;
 import com.transformuk.hee.tis.genericupload.api.dto.PersonUpdateXls;
 import com.transformuk.hee.tis.genericupload.api.dto.PersonXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.PlacementDeleteXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.PlacementUpdateXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.PlacementXls;
 import com.transformuk.hee.tis.genericupload.api.dto.PostCreateXls;
-import com.transformuk.hee.tis.genericupload.api.dto.PostFundingUpdateXLS;
+import com.transformuk.hee.tis.genericupload.api.dto.PostFundingCreateRow;
 import com.transformuk.hee.tis.genericupload.api.dto.PostUpdateXLS;
 import com.transformuk.hee.tis.genericupload.api.dto.ProgrammeMembershipUpdateXls;
 import com.transformuk.hee.tis.genericupload.api.dto.TemplateXLS;
@@ -81,9 +81,9 @@ public class ScheduledUploadTask {
   @Autowired
   private PostUpdateTransformerService postUpdateTransformerService;
   @Autowired
-  private PostFundingUpdateTransformerService postFundingUpdateTransformerService;
+  private PostFundingCreateTransformerService postFundingCreateTransformerService;
   @Autowired
-  private FundingUpdateTransformerService fundingUpdateTransformerService;
+  private PostFundingUpdateTransformerService postFundingUpdateTransformerService;
   @Autowired
   private AssessmentUpdateTransformerService assessmentUpdateTransformerService;
   @Autowired
@@ -194,21 +194,20 @@ public class ScheduledUploadTask {
             setJobToCompleted(applicationType, postUpdateXLSList);
             break;
 
-          case POSTS_FUNDING_UPDATE:
-            List<PostFundingUpdateXLS> postFundingUpdateXlsList = excelToObjectMapper
-                .map(PostFundingUpdateXLS.class,
-                    new ColumnMapper(PostFundingUpdateXLS.class).getFieldMap());
-            postFundingUpdateTransformerService
-                .processPostFundingUpdateUpload(postFundingUpdateXlsList);
-            setJobToCompleted(applicationType, postFundingUpdateXlsList);
+          case POSTS_FUNDING_CREATE:
+            List<PostFundingCreateRow> postFundingCreateRows = excelToObjectMapper
+                .map(PostFundingCreateRow.class,
+                    new ColumnMapper(PostFundingCreateRow.class).getFieldMap());
+            postFundingCreateTransformerService.processRows(postFundingCreateRows);
+            setJobToCompleted(applicationType, postFundingCreateRows);
             break;
 
-          case FUNDING_UPDATE:
-            List<FundingUpdateXLS> fundingUpdateXLSList = excelToObjectMapper
-                .map(FundingUpdateXLS.class,
-                    new ColumnMapper(FundingUpdateXLS.class).getFieldMap());
-            fundingUpdateTransformerService.processFundingUpdateUpload(fundingUpdateXLSList);
-            setJobToCompleted(applicationType, fundingUpdateXLSList);
+          case POSTS_FUNDING_UPDATE:
+            List<PostFundingUpdateRow> postFundingUpdateRows = excelToObjectMapper
+                .map(PostFundingUpdateRow.class,
+                    new ColumnMapper(PostFundingUpdateRow.class).getFieldMap());
+            postFundingUpdateTransformerService.processRows(postFundingUpdateRows);
+            setJobToCompleted(applicationType, postFundingUpdateRows);
             break;
 
           case ASSESSMENTS_UPDATE:
